Flatten directories
