+++ hosts/lookup.go
