+++ plugins/gce/gce.go
@@ -97,7 +97,7 @@ func AuthUrl() string {
 func LoadGCEConfig() *oauth.Config {
     var fileName string
     if _, err := os.Stat("/config/plugins/gce.json"); os.IsNotExist(err) {
-        fileName = "../config/plugins/gce.json"
+        fileName = "./config/plugins/gce.json"
     } else {
         fileName = "/config/plugins/gce.json"
     }
